requirements.txt
openai
python-dotenv
selenium
pytest